\name{eqs2lavaan-package}
\alias{eqs2lavaan-package}
\docType{package}
\title{
EQS Output Conversion to lavaan Functions
}
\description{
Transitioning from EQS to R for structural equation modeling (SEM) is made easier with a set of functions to convert .out files into R code.  The EQS output can be converted into lavaan syntax and run in the R environment.  Other functions parse descriptive statistics and the covariance matrix from an EQS .out file.  A heat map plot of a covariance matrix is also included.
}
\details{
\tabular{ll}{
Package: \tab eqs2lavaan\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-25\cr
License: \tab GPL-2\cr
}
}
\author{
Craig M. Krebsbach
Maintainer: Craig M. Krebsbach \email{craigmk@my.uri.edu}
}
\references{
Bentler, P. M. (2008). EQS Program Manual. Encino, CA: Multivariate Software Inc.

Rosseel, Yves (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. \url{http://www.jstatsoft.org/v48/i02/}.
}
\keyword{package}
\keyword{EQS}
\keyword{lavaan}
\keyword{SEM}
\keyword{model}
\keyword{cov}